changing number of reservation to float

+++ api/models/reservations.js
@@ -14,7 +14,7 @@ module.exports = (sequelize, DataTypes)=>{
         transaction_end_date : { type : Sequelize.TEXT, allowNull : true },
         transaction_start_time : { type : Sequelize.TEXT, allowNull : false },
         transaction_end_time : { type : Sequelize.TEXT, allowNull : true },        
-        number_activity_reserved : { type : Sequelize.INTEGER, allowNull : false },
+        number_activity_reserved : { type : Sequelize.FLOAT, allowNull : false },
         nbr_in_party : { type : Sequelize.TEXT, allowNull : false },
         nbr_in_adult : { type : Sequelize.TEXT, allowNull : false },
         nbr_children : { type : Sequelize.TEXT, allowNull : false },
@@ -26,7 +26,7 @@ module.exports = (sequelize, DataTypes)=>{
         tableName: 'reservations'
       });
     
-      //sequelize.sync()
+    //sequelize.sync()
 
     return reservations
 }
